+++ lib/_internal/file_picker_web.dart
@@ -1,10 +1,11 @@
 import 'dart:async';
 import 'dart:js_interop';
-import 'package:web/web.dart';
 import 'dart:typed_data';
 
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
+import 'package:path/path.dart' as p;
+import 'package:web/web.dart';
 
 class FilePickerWeb extends FilePicker {
   late Element _target;
@@ -42,20 +43,20 @@ class FilePickerWeb extends FilePicker {
     List<String>? allowedExtensions,
     bool allowMultiple = false,
     Function(FilePickerStatus)? onFileLoading,
-    bool allowCompression = true,
+    @Deprecated('allowCompression is deprecated and has no effect. Use compressionQuality instead.')
+    bool allowCompression = false,
     bool withData = true,
     bool withReadStream = false,
     bool lockParentWindow = false,
     bool readSequential = false,
-    int compressionQuality = 20,
+    int compressionQuality = 0,
   }) async {
-    if (type != FileType.custom && (allowedExtensions?.isNotEmpty ?? false)) {
+    /*if (type != FileType.custom && (allowedExtensions?.isNotEmpty ?? false)) {
       throw Exception(
           'You are setting a type [$type]. Custom extension filters are only allowed with FileType.custom, please change it or remove filters.');
-    }
+    }*/
 
-    final Completer<List<PlatformFile>?> filesCompleter =
-        Completer<List<PlatformFile>?>();
+    final Completer<List<PlatformFile>?> filesCompleter = Completer<List<PlatformFile>?>();
 
     String accept = _fileType(type, allowedExtensions);
     HTMLInputElement uploadInput = HTMLInputElement();
@@ -86,9 +87,15 @@ class FilePickerWeb extends FilePicker {
         String? path,
         Stream<List<int>>? readStream,
       ) {
+        String? blobUrl;
+        if (bytes != null && bytes.isNotEmpty) {
+          final blob = Blob([bytes.toJS].toJS, BlobPropertyBag(type: file.type));
+
+          blobUrl = URL.createObjectURL(blob);
+        }
         pickedFiles.add(PlatformFile(
           name: file.name,
-          path: path,
+          path: path ?? blobUrl,
           size: bytes != null ? bytes.length : file.size,
           bytes: bytes,
           readStream: readStream,
@@ -178,6 +185,49 @@ class FilePickerWeb extends FilePicker {
     return files == null ? null : FilePickerResult(files);
   }
 
+  @override
+  Future<String?> saveFile({
+    String? dialogTitle,
+    String? fileName,
+    String? initialDirectory,
+    FileType type = FileType.any,
+    List<String>? allowedExtensions,
+    Uint8List? bytes,
+    bool lockParentWindow = false,
+  }) async {
+    if (bytes == null || bytes.isEmpty) {
+      throw ArgumentError(
+        'The bytes are required when saving a file on the web.',
+      );
+    }
+
+    if (fileName == null || fileName.isEmpty) {
+      throw ArgumentError(
+        'A file name is required when saving a file on the web.',
+      );
+    }
+
+    if (p.extension(fileName).isEmpty) {
+      throw ArgumentError(
+        'The file name should include a valid file extension.',
+      );
+    }
+
+    final blob = Blob([bytes.toJS].toJS);
+    final url = URL.createObjectURL(blob);
+
+    // Start a download by using a click event on an anchor element that contains the Blob.
+    HTMLAnchorElement()
+      ..href = url
+      ..target = 'blank' // Always open the file in a new tab.
+      ..download = fileName
+      ..click();
+
+    // Release the Blob URL after the download started.
+    URL.revokeObjectURL(url);
+    return null;
+  }
+
   static String _fileType(FileType type, List<String>? allowedExtensions) {
     switch (type) {
       case FileType.any:
@@ -206,9 +256,8 @@ class FilePickerWeb extends FilePicker {
 
     int start = 0;
     while (start < file.size) {
-      final end = start + _readStreamChunkSize > file.size
-          ? file.size
-          : start + _readStreamChunkSize;
+      final end =
+          start + _readStreamChunkSize > file.size ? file.size : start + _readStreamChunkSize;
       final blob = file.slice(start, end);
       reader.readAsArrayBuffer(blob);
       await EventStreamProviders.loadEvent.forTarget(reader).first;